corrección de bugs

+++ src/app/components/solicitud-ayuda/solicitud-ayuda.component/solicitud-ayuda.component.ts
@@ -1,45 +1,44 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from '@angular/core';
 
 import { Solicitud } from '../../../models/solicitud.class';
 import { SolicitudService } from '../../../services/solicitud-ayuda.service';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
+import { RouterLink } from '@angular/router';
 
 @Component({
   selector: 'app-solicitud-ayuda',
   templateUrl: './solicitud-ayuda.component.html',
   styleUrls: ['./solicitud-ayuda.component.css'],
-  imports: [CommonModule, FormsModule],
+  imports: [CommonModule, FormsModule, RouterLink],
   standalone: true
 })
 
-export class SolicitudAyudaComponent implements OnInit{
+export class SolicitudAyudaComponent implements OnInit {
 
-  solicitudes:Solicitud[]=[];
+  solicitudes: Solicitud[] = [];
 
-  solicitud:Solicitud=new Solicitud();
+  solicitud: Solicitud = new Solicitud();
 
-  constructor(
-    private solicitudService:SolicitudService
-  ){}
+  private solicitudService = inject(SolicitudService);
 
-  ngOnInit(){
+  ngOnInit() {
 
     this.cargarSolicitudes();
 
   }
 
-  cargarSolicitudes(){
+  cargarSolicitudes() {
 
     this.solicitudService.getSolicitudes().subscribe({
 
-      next:(result)=>{
+      next: (result) => {
 
-        this.solicitudes=result.data;
+        this.solicitudes = result.data;
 
       },
 
-      error:(err)=>{
+      error: (err) => {
 
         console.log(err);
 
@@ -48,30 +47,25 @@ export class SolicitudAyudaComponent implements OnInit{
     });
 
   }
-    guardar() {
-      console.log('Solicitud enviada:', this.solicitud);
-      this.solicitudService.createSolicitud(this.solicitud).subscribe({
-        next: (result) => {
-          alert(result.msg);
-          this.solicitud = new Solicitud();
-          this.cargarSolicitudes();
-        },
-        error: (err) => console.error('Error al crear solicitud', err)
-      });
-    }
-
-  eliminar(id:number){
-
-    this.solicitudService.deleteSolicitud(id).subscribe({
-
-      next:()=>{
-
+  guardar() {
+    console.log('Solicitud enviada:', this.solicitud);
+    this.solicitudService.createSolicitud(this.solicitud).subscribe({
+      next: (result) => {
+        alert(result.msg);
+        this.solicitud = new Solicitud();
         this.cargarSolicitudes();
-
-      }
-
+      },
+      error: (err) => console.error('Error al crear solicitud', err)
     });
+  }
 
+  eliminar(id: number | undefined) {
+    if (id === undefined) return;
+    this.solicitudService.deleteSolicitud(id).subscribe({
+      next: () => {
+        this.cargarSolicitudes();
+      }
+    });
   }
 
 }
